Add Vue for data binding
I thought this would be a good oppourtunity to try out a JS
framework. Vue looked fairly easy to implement and after a
few issues finally got it working. So far I'm very pleased
with the outcome. The frontend didn't have much yet so swapping
in Vue isn't too bad.