Implemented device types, implement job manager and workers
Brand and Connection have been removed from the device table and model
and replaced with an ID (or instance of) a device Type.

INCA can now actually run jobs! Still no management of types, but
an "archive job" can be ran that will run scripts in the script dir
to generate a log file. An entry is added to the database and shown
on the web interface.