Removed Vue, PUT only part of a device
I decided to ditch Vue on the client side. It was added unneeded
complexity for the size of this application and was more of a
frustration and bloat than anything. Everything I was to do client
side can be easily and cleanly done with vanilla Javascript and
jQuery.

The PUT and POST endpoints for the device controller have been
separated and modified. The PUT method no longer requires a full
device struct in "compliance" with what a PUT request normally means.

Both methods have been altered to operate on a separate device object
and copy over data as needed. This provides better security and more
control regarding how devices are edited/created.