Added job status
The current status of running jobs is show on the dashboard using
the progress bar that has been empty this whole time. The client will
periodically ask the server for the current status and update the
bar accordingly.

The job manager is setup to utilize workers running in goroutines. At
minimum, each job will spawn 3 goroutines. One from the manager watching
for when the worker is done. One from the worker for its actual run
method. And one for a script execution. Multiple goroutines may be
spawned depending on how many devices are included in the job. Currently
there's no limit on the number of running jobs/workers. Eventually this
is something that needs to be added but it's not critical at this time.