Renamed Types to Connection Profiles
Connection Profile is a better name for what a device type used to be.
In the new system, there's now need for "brand" and "connection" as
a device is assigned a specific profile with that information.
Brand and Connection have been removed as they're not needed.
Connection profiles hold a remote username/password combo and an
enable password for Cisco devices.

Scripts are no longer executed with arguments. Instead, all variables
are placed into the environment for the script. All Inca vars will
be prefixed with INCA_. So far the following variables exist:

INCA_ADDRESS
INCA_REMOTE_USERNAME
INCA_REMOTE_PASSWORD
INCA_OUTPUT_FILE
INCA_ENABLE_PASSWORD

An executable may use them as needed.