Dist Make task, proper relative URLs
Added a Make task named "dist" which will create a distributable
tarball of Inca.

A new configuration option was added Core.SiteBasePath. This should
be set if Inca is running under a "sub directory" within proxy or
webserver like Nginx or Apache. URLs must be rewritten to remove the
subdirectory so Inca sees it as if it were the root path.