Created tool to sync the database with configs on the filesystem
This tool will scan the configs directory and check if the config
is in the database. If it's not, it will be added if the device
exists. The filesystem is kept in sync with the database each time
a device page is loaded so there's no need for a tool that goes
the other way.