Fixed database issues when getting configs
For high volume devices, many configs were erroring with "database
is locked". This has been mitigated by having the save piece in a
separate go routine and configs are passed to it once they're ready
to be saved. This way the gathering is concurrent, but the saving
is linear to reduce errors.

Currently it's slower than before but can be sped up by using bulk
inserts. That will come in a later update.